requirements.txt
crewai>=0.11.0
langchain>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=0.19.0
transformers>=4.30.0
torch>=2.0.0
fastapi[all]>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
python-multipart>=0.0.6 